Moved event classes to events package
