+++ common/src/main/java/org/mcaccess/minecraftaccess/features/NarrateCrosshair.java
@@ -27,7 +27,7 @@ import org.jetbrains.annotations.Nullable;
 import org.mcaccess.minecraftaccess.Config;
 import org.mcaccess.minecraftaccess.MainClass;
 import org.mcaccess.minecraftaccess.api.WorldNarrator;
-import org.mcaccess.minecraftaccess.utils.ClientPlayingTick;
+import org.mcaccess.minecraftaccess.utils.events.ClientPlayingTick;
 import org.mcaccess.minecraftaccess.utils.condition.Interval;
 
 /**
