+++ common/src/main/java/org/mcaccess/minecraftaccess/features/NarrateHeldItem.java
@@ -22,7 +22,7 @@ import org.jetbrains.annotations.NotNull;
 
 import org.mcaccess.minecraftaccess.Config;
 import org.mcaccess.minecraftaccess.MainClass;
-import org.mcaccess.minecraftaccess.utils.ClientPlayingTick;
+import org.mcaccess.minecraftaccess.utils.events.ClientPlayingTick;
 import org.mcaccess.minecraftaccess.utils.KeyMappingCategories;
 
 public class NarrateHeldItem implements BalmClientModule {
