+++ common/src/main/java/org/mcaccess/minecraftaccess/features/point_of_interest/ObjectTracker.java
@@ -29,9 +29,9 @@ import org.jetbrains.annotations.NotNull;
 import org.mcaccess.minecraftaccess.Config;
 import org.mcaccess.minecraftaccess.MainClass;
 import org.mcaccess.minecraftaccess.api.WorldNarrator;
-import org.mcaccess.minecraftaccess.utils.ClientPlayingTick;
 import org.mcaccess.minecraftaccess.utils.KeyMappingCategories;
 import org.mcaccess.minecraftaccess.utils.NarrationUtils;
+import org.mcaccess.minecraftaccess.utils.events.ClientPlayingTick;
 
 public class ObjectTracker implements BalmClientModule {
     public static final String START_OF_LIST = "minecraft_access.other.start_of_list";
