+++ common/src/main/java/org/mcaccess/minecraftaccess/features/point_of_interest/POIBlocks.java
@@ -22,7 +22,7 @@ import org.jetbrains.annotations.Nullable;
 
 import org.mcaccess.minecraftaccess.Config;
 import org.mcaccess.minecraftaccess.MainClass;
-import org.mcaccess.minecraftaccess.utils.ClientPlayingTick;
+import org.mcaccess.minecraftaccess.utils.events.ClientPlayingTick;
 import org.mcaccess.minecraftaccess.utils.condition.Interval;
 
 /**
